# backend/swagger.json
{"swagger":"2.0","info":{"description":"","title":"standard public schema","version":"13.0.5"},"host":"nzwmoewutowpkdlzzgty.supabase.co:443","basePath":"/","schemes":["https"],"consumes":["application/json","application/vnd.pgrst.object+json;nulls=stripped","application/vnd.pgrst.object+json","text/csv"],"produces":["application/json","application/vnd.pgrst.object+json;nulls=stripped","application/vnd.pgrst.object+json","text/csv"],"paths":{"/":{"get":{"produces":["application/openapi+json","application/json"],"responses":{"200":{"description":"OK"}},"summary":"OpenAPI description (this document)","tags":["Introspection"]}},"/letters":{"get":{"parameters":[{"$ref":"#/parameters/rowFilter.letters.letter_id"},{"$ref":"#/parameters/rowFilter.letters.user_id"},{"$ref":"#/parameters/rowFilter.letters.contract_id"},{"$ref":"#/parameters/rowFilter.letters.letter_type"},{"$ref":"#/parameters/rowFilter.letters.generated_content"},{"$ref":"#/parameters/rowFilter.letters.customizations"},{"$ref":"#/parameters/rowFilter.letters.created_at"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"200":{"description":"OK","schema":{"items":{"$ref":"#/definitions/letters"},"type":"array"}},"206":{"description":"Partial Content"}},"tags":["letters"]},"post":{"parameters":[{"$ref":"#/parameters/body.letters"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/preferPost"}],"responses":{"201":{"description":"Created"}},"tags":["letters"]},"delete":{"parameters":[{"$ref":"#/parameters/rowFilter.letters.letter_id"},{"$ref":"#/parameters/rowFilter.letters.user_id"},{"$ref":"#/parameters/rowFilter.letters.contract_id"},{"$ref":"#/parameters/rowFilter.letters.letter_type"},{"$ref":"#/parameters/rowFilter.letters.generated_content"},{"$ref":"#/parameters/rowFilter.letters.customizations"},{"$ref":"#/parameters/rowFilter.letters.created_at"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["letters"]},"patch":{"parameters":[{"$ref":"#/parameters/rowFilter.letters.letter_id"},{"$ref":"#/parameters/rowFilter.letters.user_id"},{"$ref":"#/parameters/rowFilter.letters.contract_id"},{"$ref":"#/parameters/rowFilter.letters.letter_type"},{"$ref":"#/parameters/rowFilter.letters.generated_content"},{"$ref":"#/parameters/rowFilter.letters.customizations"},{"$ref":"#/parameters/rowFilter.letters.created_at"},{"$ref":"#/parameters/body.letters"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["letters"]}},"/advice":{"get":{"parameters":[{"$ref":"#/parameters/rowFilter.advice.id"},{"$ref":"#/parameters/rowFilter.advice.contract_type"},{"$ref":"#/parameters/rowFilter.advice.section"},{"$ref":"#/parameters/rowFilter.advice.title"},{"$ref":"#/parameters/rowFilter.advice.content"},{"$ref":"#/parameters/rowFilter.advice.icon"},{"$ref":"#/parameters/rowFilter.advice.created_at"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"200":{"description":"OK","schema":{"items":{"$ref":"#/definitions/advice"},"type":"array"}},"206":{"description":"Partial Content"}},"tags":["advice"]},"post":{"parameters":[{"$ref":"#/parameters/body.advice"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/preferPost"}],"responses":{"201":{"description":"Created"}},"tags":["advice"]},"delete":{"parameters":[{"$ref":"#/parameters/rowFilter.advice.id"},{"$ref":"#/parameters/rowFilter.advice.contract_type"},{"$ref":"#/parameters/rowFilter.advice.section"},{"$ref":"#/parameters/rowFilter.advice.title"},{"$ref":"#/parameters/rowFilter.advice.content"},{"$ref":"#/parameters/rowFilter.advice.icon"},{"$ref":"#/parameters/rowFilter.advice.created_at"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["advice"]},"patch":{"parameters":[{"$ref":"#/parameters/rowFilter.advice.id"},{"$ref":"#/parameters/rowFilter.advice.contract_type"},{"$ref":"#/parameters/rowFilter.advice.section"},{"$ref":"#/parameters/rowFilter.advice.title"},{"$ref":"#/parameters/rowFilter.advice.content"},{"$ref":"#/parameters/rowFilter.advice.icon"},{"$ref":"#/parameters/rowFilter.advice.created_at"},{"$ref":"#/parameters/body.advice"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["advice"]}},"/users":{"get":{"parameters":[{"$ref":"#/parameters/rowFilter.users.user_id"},{"$ref":"#/parameters/rowFilter.users.email"},{"$ref":"#/parameters/rowFilter.users.password_hash"},{"$ref":"#/parameters/rowFilter.users.subscription_tier"},{"$ref":"#/parameters/rowFilter.users.subscription_expiry"},{"$ref":"#/parameters/rowFilter.users.scan_count"},{"$ref":"#/parameters/rowFilter.users.stripe_customer_id"},{"$ref":"#/parameters/rowFilter.users.created_at"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"200":{"description":"OK","schema":{"items":{"$ref":"#/definitions/users"},"type":"array"}},"206":{"description":"Partial Content"}},"tags":["users"]},"post":{"parameters":[{"$ref":"#/parameters/body.users"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/preferPost"}],"responses":{"201":{"description":"Created"}},"tags":["users"]},"delete":{"parameters":[{"$ref":"#/parameters/rowFilter.users.user_id"},{"$ref":"#/parameters/rowFilter.users.email"},{"$ref":"#/parameters/rowFilter.users.password_hash"},{"$ref":"#/parameters/rowFilter.users.subscription_tier"},{"$ref":"#/parameters/rowFilter.users.subscription_expiry"},{"$ref":"#/parameters/rowFilter.users.scan_count"},{"$ref":"#/parameters/rowFilter.users.stripe_customer_id"},{"$ref":"#/parameters/rowFilter.users.created_at"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["users"]},"patch":{"parameters":[{"$ref":"#/parameters/rowFilter.users.user_id"},{"$ref":"#/parameters/rowFilter.users.email"},{"$ref":"#/parameters/rowFilter.users.password_hash"},{"$ref":"#/parameters/rowFilter.users.subscription_tier"},{"$ref":"#/parameters/rowFilter.users.subscription_expiry"},{"$ref":"#/parameters/rowFilter.users.scan_count"},{"$ref":"#/parameters/rowFilter.users.stripe_customer_id"},{"$ref":"#/parameters/rowFilter.users.created_at"},{"$ref":"#/parameters/body.users"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["users"]}},"/analyses":{"get":{"parameters":[{"$ref":"#/parameters/rowFilter.analyses.id"},{"$ref":"#/parameters/rowFilter.analyses.user_id"},{"$ref":"#/parameters/rowFilter.analyses.source"},{"$ref":"#/parameters/rowFilter.analyses.source_url"},{"$ref":"#/parameters/rowFilter.analyses.full_text"},{"$ref":"#/parameters/rowFilter.analyses.summary"},{"$ref":"#/parameters/rowFilter.analyses.flags"},{"$ref":"#/parameters/rowFilter.analyses.bounding_boxes"},{"$ref":"#/parameters/rowFilter.analyses.ocr_confidence"},{"$ref":"#/parameters/rowFilter.analyses.rules_version"},{"$ref":"#/parameters/rowFilter.analyses.llm_model"},{"$ref":"#/parameters/rowFilter.analyses.metadata"},{"$ref":"#/parameters/rowFilter.analyses.created_at"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"200":{"description":"OK","schema":{"items":{"$ref":"#/definitions/analyses"},"type":"array"}},"206":{"description":"Partial Content"}},"tags":["analyses"]},"post":{"parameters":[{"$ref":"#/parameters/body.analyses"},{"$ref":"#/parameters/select"},{"$ref":"#/parameters/preferPost"}],"responses":{"201":{"description":"Created"}},"tags":["analyses"]},"delete":{"parameters":[{"$ref":"#/parameters/rowFilter.analyses.id"},{"$ref":"#/parameters/rowFilter.analyses.user_id"},{"$ref":"#/parameters/rowFilter.analyses.source"},{"$ref":"#/parameters/rowFilter.analyses.source_url"},{"$ref":"#/parameters/rowFilter.analyses.full_text"},{"$ref":"#/parameters/rowFilter.analyses.summary"},{"$ref":"#/parameters/rowFilter.analyses.flags"},{"$ref":"#/parameters/rowFilter.analyses.bounding_boxes"},{"$ref":"#/parameters/rowFilter.analyses.ocr_confidence"},{"$ref":"#/parameters/rowFilter.analyses.rules_version"},{"$ref":"#/parameters/rowFilter.analyses.llm_model"},{"$ref":"#/parameters/rowFilter.analyses.metadata"},{"$ref":"#/parameters/rowFilter.analyses.created_at"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["analyses"]},"patch":{"parameters":[{"$ref":"#/parameters/rowFilter.analyses.id"},{"$ref":"#/parameters/rowFilter.analyses.user_id"},{"$ref":"#/parameters/rowFilter.analyses.source"},{"$ref":"#/parameters/rowFilter.analyses.source_url"},{"$ref":"#/parameters/rowFilter.analyses.full_text"},{"$ref":"#/parameters/rowFilter.analyses.summary"},{"$ref":"#/parameters/rowFilter.analyses.flags"},{"$ref":"#/parameters/rowFilter.analyses.bounding_boxes"},{"$ref":"#/parameters/rowFilter.analyses.ocr_confidence"},{"$ref":"#/parameters/rowFilter.analyses.rules_version"},{"$ref":"#/parameters/rowFilter.analyses.llm_model"},{"$ref":"#/parameters/rowFilter.analyses.metadata"},{"$ref":"#/parameters/rowFilter.analyses.created_at"},{"$ref":"#/parameters/body.analyses"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}},"tags":["analyses"]}}},"definitions":{"letters":{"required":["letter_id","user_id","letter_type"],"properties":{"letter_id":{"default":"extensions.uuid_generate_v4()","description":"Note:\nThis is a Primary Key.<pk/>","format":"uuid","type":"string"},"user_id":{"description":"Note:\nThis is a Foreign Key to `users.user_id`.<fk table='users' column='user_id'/>","format":"uuid","type":"string"},"contract_id":{"description":"Note:\nThis is a Foreign Key to `analyses.id`.<fk table='analyses' column='id'/>","format":"uuid","type":"string"},"letter_type":{"format":"text","type":"string"},"generated_content":{"format":"text","type":"string"},"customizations":{"format":"jsonb"},"created_at":{"default":"now()","format":"timestamp with time zone","type":"string"}},"type":"object"},"advice":{"required":["id","contract_type","section","title","content"],"properties":{"id":{"default":"extensions.uuid_generate_v4()","description":"Note:\nThis is a Primary Key.<pk/>","format":"uuid","type":"string"},"contract_type":{"format":"text","type":"string"},"section":{"format":"text","type":"string"},"title":{"format":"text","type":"string"},"content":{"format":"text","type":"string"},"icon":{"format":"text","type":"string"},"created_at":{"default":"now()","format":"timestamp with time zone","type":"string"}},"type":"object"},"users":{"required":["user_id","email","password_hash"],"properties":{"user_id":{"default":"extensions.uuid_generate_v4()","description":"Note:\nThis is a Primary Key.<pk/>","format":"uuid","type":"string"},"email":{"format":"text","type":"string"},"password_hash":{"format":"text","type":"string"},"subscription_tier":{"default":"free","format":"text","type":"string"},"subscription_expiry":{"format":"timestamp with time zone","type":"string"},"scan_count":{"default":0,"format":"integer","type":"integer"},"stripe_customer_id":{"format":"text","type":"string"},"created_at":{"default":"now()","format":"timestamp with time zone","type":"string"}},"type":"object"},"analyses":{"required":["id"],"properties":{"id":{"default":"extensions.uuid_generate_v4()","description":"Note:\nThis is a Primary Key.<pk/>","format":"uuid","type":"string"},"user_id":{"description":"Note:\nThis is a Foreign Key to `users.user_id`.<fk table='users' column='user_id'/>","format":"uuid","type":"string"},"source":{"default":"file","format":"text","type":"string"},"source_url":{"format":"text","type":"string"},"full_text":{"format":"text","type":"string"},"summary":{"format":"text","type":"string"},"flags":{"format":"jsonb"},"bounding_boxes":{"format":"jsonb"},"ocr_confidence":{"format":"double precision","type":"number"},"rules_version":{"format":"text","type":"string"},"llm_model":{"format":"text","type":"string"},"metadata":{"format":"jsonb"},"created_at":{"default":"now()","format":"timestamp with time zone","type":"string"}},"type":"object"}},"parameters":{"preferParams":{"name":"Prefer","description":"Preference","required":false,"enum":[],"in":"header","type":"string"},"preferReturn":{"name":"Prefer","description":"Preference","required":false,"enum":["return=representation","return=minimal","return=none"],"in":"header","type":"string"},"preferCount":{"name":"Prefer","description":"Preference","required":false,"enum":["count=none"],"in":"header","type":"string"},"preferPost":{"name":"Prefer","description":"Preference","required":false,"enum":["return=representation","return=minimal","return=none","resolution=ignore-duplicates","resolution=merge-duplicates"],"in":"header","type":"string"},"select":{"name":"select","description":"Filtering Columns","required":false,"in":"query","type":"string"},"on_conflict":{"name":"on_conflict","description":"On Conflict","required":false,"in":"query","type":"string"},"order":{"name":"order","description":"Ordering","required":false,"in":"query","type":"string"},"range":{"name":"Range","description":"Limiting and Pagination","required":false,"in":"header","type":"string"},"rangeUnit":{"name":"Range-Unit","description":"Limiting and Pagination","required":false,"default":"items","in":"header","type":"string"},"offset":{"name":"offset","description":"Limiting and Pagination","required":false,"in":"query","type":"string"},"limit":{"name":"limit","description":"Limiting and Pagination","required":false,"in":"query","type":"string"},"body.letters":{"name":"letters","description":"letters","required":false,"in":"body","schema":{"$ref":"#/definitions/letters"}},"rowFilter.letters.letter_id":{"name":"letter_id","required":false,"in":"query","type":"string"},"rowFilter.letters.user_id":{"name":"user_id","required":false,"in":"query","type":"string"},"rowFilter.letters.contract_id":{"name":"contract_id","required":false,"in":"query","type":"string"},"rowFilter.letters.letter_type":{"name":"letter_type","required":false,"in":"query","type":"string"},"rowFilter.letters.generated_content":{"name":"generated_content","required":false,"in":"query","type":"string"},"rowFilter.letters.customizations":{"name":"customizations","required":false,"in":"query","type":"string"},"rowFilter.letters.created_at":{"name":"created_at","required":false,"in":"query","type":"string"},"body.advice":{"name":"advice","description":"advice","required":false,"in":"body","schema":{"$ref":"#/definitions/advice"}},"rowFilter.advice.id":{"name":"id","required":false,"in":"query","type":"string"},"rowFilter.advice.contract_type":{"name":"contract_type","required":false,"in":"query","type":"string"},"rowFilter.advice.section":{"name":"section","required":false,"in":"query","type":"string"},"rowFilter.advice.title":{"name":"title","required":false,"in":"query","type":"string"},"rowFilter.advice.content":{"name":"content","required":false,"in":"query","type":"string"},"rowFilter.advice.icon":{"name":"icon","required":false,"in":"query","type":"string"},"rowFilter.advice.created_at":{"name":"created_at","required":false,"in":"query","type":"string"},"body.users":{"name":"users","description":"users","required":false,"in":"body","schema":{"$ref":"#/definitions/users"}},"rowFilter.users.user_id":{"name":"user_id","required":false,"in":"query","type":"string"},"rowFilter.users.email":{"name":"email","required":false,"in":"query","type":"string"},"rowFilter.users.password_hash":{"name":"password_hash","required":false,"in":"query","type":"string"},"rowFilter.users.subscription_tier":{"name":"subscription_tier","required":false,"in":"query","type":"string"},"rowFilter.users.subscription_expiry":{"name":"subscription_expiry","required":false,"in":"query","type":"string"},"rowFilter.users.scan_count":{"name":"scan_count","required":false,"in":"query","type":"string"},"rowFilter.users.stripe_customer_id":{"name":"stripe_customer_id","required":false,"in":"query","type":"string"},"rowFilter.users.created_at":{"name":"created_at","required":false,"in":"query","type":"string"},"body.analyses":{"name":"analyses","description":"analyses","required":false,"in":"body","schema":{"$ref":"#/definitions/analyses"}},"rowFilter.analyses.id":{"name":"id","required":false,"in":"query","type":"string"},"rowFilter.analyses.user_id":{"name":"user_id","required":false,"in":"query","type":"string"},"rowFilter.analyses.source":{"name":"source","required":false,"in":"query","type":"string"},"rowFilter.analyses.source_url":{"name":"source_url","required":false,"in":"query","type":"string"},"rowFilter.analyses.full_text":{"name":"full_text","required":false,"in":"query","type":"string"},"rowFilter.analyses.summary":{"name":"summary","required":false,"in":"query","type":"string"},"rowFilter.analyses.flags":{"name":"flags","required":false,"in":"query","type":"string"},"rowFilter.analyses.bounding_boxes":{"name":"bounding_boxes","required":false,"in":"query","type":"string"},"rowFilter.analyses.ocr_confidence":{"name":"ocr_confidence","required":false,"in":"query","type":"string"},"rowFilter.analyses.rules_version":{"name":"rules_version","required":false,"in":"query","type":"string"},"rowFilter.analyses.llm_model":{"name":"llm_model","required":false,"in":"query","type":"string"},"rowFilter.analyses.metadata":{"name":"metadata","required":false,"in":"query","type":"string"},"rowFilter.analyses.created_at":{"name":"created_at","required":false,"in":"query","type":"string"}},"externalDocs":{"description":"PostgREST Documentation","url":"https://postgrest.org/en/v13/references/api.html"}}